chore(numbertext): rename NumberLockUps to NumberText
