+++ src/components/molecules/NumberText/NumberText.test.tsx
@@ -2,13 +2,13 @@ import { axe } from 'jest-axe';
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 
-import NumberLockUps from './NumberLockUps';
+import NumberText from './NumberText';
 
 import 'jest-styled-components';
 
-const renderComponent = (props = {}) => render(<NumberLockUps title="Overall Balance" value={100000} {...props} />);
+const renderComponent = (props = {}) => render(<NumberText title="Overall Balance" value={100000} {...props} />);
 
-describe('<NumberLockUps />', () => {
+describe('<NumberText />', () => {
   it('renders without errors', async () => {
     const { container } = renderComponent({
       formatterOptions: {
